Mapeamento das entidades

+++ ecommerce.API/Ecommerce.Bussines/Models/Pedido.cs
@@ -5,7 +5,7 @@ namespace Ecommerce.Bussines.Models
     public class Pedido
     {
         public int Id { get; set; }
-        public double TotalPedido { get; set; }
+        public decimal TotalPedido { get; set; }
         public Pagamento Pagamento { get; set; }
         public  Cliente Cliente { get; set; }
         public IEnumerable<ItemPedido> ItemsPedidos { get; set; }
